requirements.txt
pandas
streamlit==1.11
bokeh
plotly
# plotly-express
st_pages==0.4.4
# streamlit_extras==0.3.6
matplotlib
scikit-learn
numpy
streamlit_plotly_events
seaborn
pybaselines
# altair<5
hydra-core
omegaconf
torchmetrics
hydra_colorlog